Bugfix/fix factorybot ci issue (#583)
* Add new constraints for several gemfiles

* Update readme

* Ensure generated app has lower factory bot

* Split helpers into 2 files

* Update helper doc

* Add changelog
