+++ Appraisals
@@ -4,9 +4,10 @@ appraise 'rails_5_2' do
   gem 'activerecord'
   gem 'capybara', '< 3.33'
   gem 'cucumber', '< 6'
+  gem 'factory_bot', '< 6.4'
   gem 'psych', '< 4'
   gem 'rails-html-sanitizer', '< 1.4.3'
-  gem 'railties', '~> 5.2.4'
+  gem 'railties', '~> 5.2.8'
   gem 'sqlite3', '~> 1.3.13'
 end
 
@@ -14,30 +15,32 @@ appraise 'rails_6_0' do
   gem 'activerecord'
   gem 'capybara', '< 3.35'
   gem 'cucumber', '< 6'
-  gem 'matrix' # Until capybara 3.36+
+  gem 'factory_bot', '< 6.4'
+  gem 'matrix'
   gem 'psych', '< 4'
   gem 'rails-html-sanitizer', '< 1.4.3'
-  gem 'railties', '~> 6.0.3'
-  gem 'sqlite3', '~> 1.4'
+  gem 'railties', '~> 6.0.6'
+  gem 'sqlite3', '< 1.6'
 end
 
 appraise 'rails_6_1' do
   gem 'activerecord'
   gem 'capybara', '< 3.38'
+  gem 'factory_bot', '< 6.4'
   gem 'psych', '< 4'
-  gem 'railties', '~> 6.1.3'
+  gem 'railties', '~> 6.1.7'
   gem 'sqlite3', '~> 1.4'
 end
 
 appraise 'rails_7_0' do
   gem 'activerecord'
   gem 'cucumber', '< 10'
-  gem 'railties', '~> 7.0.0'
-  gem 'sqlite3', '~> 1.4'
+  gem 'railties', '~> 7.0.8'
+  gem 'sqlite3', '~> 1.7'
 end
 
 appraise 'rails_7_1' do
   gem 'activerecord'
-  gem 'railties',    '~> 7.1.0'
-  gem 'sqlite3', '~> 1.4'
+  gem 'railties', '~> 7.1.3'
+  gem 'sqlite3', '~> 1.7'
 end
