+++ gemfiles/rails_5_2.gemfile
@@ -5,9 +5,10 @@ source "https://rubygems.org"
 gem "activerecord"
 gem "capybara", "< 3.33"
 gem "cucumber", "< 6"
+gem "factory_bot", "< 6.4"
 gem "psych", "< 4"
 gem "rails-html-sanitizer", "< 1.4.3"
-gem "railties", "~> 5.2.4"
+gem "railties", "~> 5.2.8"
 gem "sqlite3", "~> 1.3.13"
 
 gemspec path: "../"
