+++ gemfiles/rails_6_0.gemfile
@@ -5,10 +5,11 @@ source "https://rubygems.org"
 gem "activerecord"
 gem "capybara", "< 3.35"
 gem "cucumber", "< 6"
+gem "factory_bot", "< 6.4"
 gem "matrix"
 gem "psych", "< 4"
 gem "rails-html-sanitizer", "< 1.4.3"
-gem "railties", "~> 6.0.3"
-gem "sqlite3", "~> 1.4"
+gem "railties", "~> 6.0.6"
+gem "sqlite3", "< 1.6"
 
 gemspec path: "../"
