+++ gemfiles/rails_6_1.gemfile
@@ -4,8 +4,9 @@ source "https://rubygems.org"
 
 gem "activerecord"
 gem "capybara", "< 3.38"
+gem "factory_bot", "< 6.4"
 gem "psych", "< 4"
-gem "railties", "~> 6.1.3"
+gem "railties", "~> 6.1.7"
 gem "sqlite3", "~> 1.4"
 
 gemspec path: "../"
