+++ gemfiles/rails_7_1.gemfile
@@ -3,7 +3,7 @@
 source "https://rubygems.org"
 
 gem "activerecord"
-gem "railties", "~> 7.1.0"
-gem "sqlite3", "~> 1.4"
+gem "railties", "~> 7.1.3"
+gem "sqlite3", "~> 1.7"
 
 gemspec path: "../"
